+++ lib/native_video_editor.dart
@@ -1,3 +1,4 @@
 export 'src/models/video_crop_rect.dart';
 export 'src/models/video_edit_request.dart';
+export 'src/models/video_thumbnail_request.dart';
 export 'src/native_video_editor.dart';